feat: add shape component hit test
