+++ packages/effects-core/src/components/mesh-component.ts
@@ -1,0 +1,46 @@
+import { serialize } from '../decorators';
+import type { BoundingBoxTriangle, HitTestTriangleParams } from '../plugins';
+import { MeshCollider } from '../plugins';
+import type { Geometry } from '../render/geometry';
+import type { Renderer } from '../render/renderer';
+import { RendererComponent } from './renderer-component';
+
+export class MeshComponent extends RendererComponent {
+  /**
+   * 渲染的 Geometry
+   */
+  @serialize()
+  protected geometry: Geometry;
+  /**
+   * 用于点击测试的碰撞器
+   */
+  protected meshCollider = new MeshCollider();
+
+  override render (renderer: Renderer) {
+    if (renderer.renderingData.currentFrame.globalUniforms) {
+      renderer.setGlobalMatrix('effects_ObjectToWorld', this.transform.getWorldMatrix());
+    }
+    renderer.drawGeometry(this.geometry, this.material);
+  }
+
+  // TODO 点击测试后续抽象一个 Collider 组件
+  getHitTestParams = (force?: boolean): HitTestTriangleParams | void => {
+    const area = this.getBoundingBox();
+
+    if (area) {
+      return {
+        type: area.type,
+        triangles: area.area,
+      };
+    }
+  };
+
+  getBoundingBox (): BoundingBoxTriangle | void {
+    const worldMatrix = this.transform.getWorldMatrix();
+
+    this.meshCollider.setGeometry(this.geometry, worldMatrix);
+    const boundingBoxData = this.meshCollider.getBoundingBoxData();
+
+    return boundingBoxData;
+  }
+}
